+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ CHANGELOG
 
 Release NEXT
 ------------
+- Nothing yet.
+
+Release 0.134.0
+---------------
 - Drop django-permission.
 - Restructure admin menu.